+++ test/src/rules/one-level/element-types.spec.js
@@ -1,0 +1,281 @@
+const { ELEMENT_TYPES: RULE } = require("../../../../src/constants/rules");
+const { SETTINGS, createRuleTester, pathResolvers } = require("../../helpers");
+
+const rule = require(`../../../../src/rules/${RULE}`);
+
+const { absoluteFilePath, codeFilePath } = pathResolvers("one-level");
+
+const errorMessage = (fileType, dependencyType) =>
+  `Usage of '${dependencyType}' is not allowed in '${fileType}'`;
+
+const test = (settings, options) => {
+  const ruleTester = createRuleTester(settings);
+
+  ruleTester.run(RULE, rule, {
+    valid: [
+      // Non recognized types can import whatever
+      {
+        filename: absoluteFilePath("foo/index.js"),
+        code: "import HelperA from 'helpers/helper-a'",
+        options,
+      },
+      // Components can import helpers
+      {
+        filename: absoluteFilePath("components/component-a/ComponentA.js"),
+        code: "import HelperA from '../../helpers/helper-a'",
+        options,
+      },
+      // Components can import helpers using alias
+      {
+        filename: absoluteFilePath("components/component-a/ComponentA.js"),
+        code: "import HelperA from 'helpers/helper-a'",
+        options,
+      },
+      // Components can import components using alias
+      {
+        filename: absoluteFilePath("components/component-a/ComponentA.js"),
+        code: "import ComponentB from 'components/component-b'",
+        options,
+      },
+      // Modules can import helpers using alias
+      {
+        filename: absoluteFilePath("modules/module-a/ModuleA.js"),
+        code: "import HelperA from 'helpers/helper-a'",
+        options,
+      },
+      // Modules can import any helpers file using alias
+      {
+        filename: absoluteFilePath("modules/module-a/ModuleA.js"),
+        code: "import HelperA from 'helpers/helper-a/HelperA.js'",
+        options,
+      },
+      // Modules can import components using alias
+      {
+        filename: absoluteFilePath("modules/module-a/ModuleA.js"),
+        code: "import ComponentA from 'components/component-a'",
+        options,
+      },
+      // Modules can import other not recognized types when alias is not set
+      {
+        filename: absoluteFilePath("modules/module-a/ModuleA.js"),
+        code: "import ComponentA from 'components/component-a'",
+        options,
+        settings: {
+          ...settings,
+          "boundaries/alias": null,
+        },
+      },
+      // Can import internal files
+      {
+        filename: absoluteFilePath("modules/module-a/index.js"),
+        code: "import ModuleB from './ModuleA'",
+        options,
+      },
+      // Modules can import modules
+      {
+        filename: absoluteFilePath("modules/module-a/ModuleA.js"),
+        code: "import ModuleB from '../module-b'",
+        options,
+      },
+      // Modules can import non existant modules files
+      {
+        filename: absoluteFilePath("modules/module-a/ModuleA.js"),
+        code: "import ModuleB from '../../modules/module-b/foo.js'",
+        options,
+      },
+      // Helpers can import ignored helpers
+      {
+        filename: absoluteFilePath("helpers/helper-a/HelperA.js"),
+        code: "import HelperB from 'helpers/helper-b'",
+        options,
+        settings: {
+          ...settings,
+          "boundaries/ignore": [codeFilePath("helpers/helper-b/**/*.js")],
+        },
+      },
+      // Invalid options
+      {
+        filename: absoluteFilePath("modules/module-a/ModuleA.js"),
+        code: "import ModuleB from '../module-b/foo.js'",
+        options: [{ rules: undefined }],
+      },
+      // Invalid options
+      {
+        filename: absoluteFilePath("modules/module-a/ModuleA.js"),
+        code: "import ModuleB from '../module-b/foo.js'",
+        options: [
+          {
+            rules: [
+              {
+                from: "foo",
+                allow: ["components"],
+              },
+            ],
+          },
+        ],
+      },
+      // Invalid options
+      {
+        filename: absoluteFilePath("modules/module-a/ModuleA.js"),
+        code: "import ModuleB from '../../modules/module-b/foo.js'",
+        options: [
+          {
+            rules: [
+              {
+                from: "components",
+                disallow: ["foo"],
+              },
+            ],
+          },
+        ],
+      },
+      // No types provided in settings
+      {
+        filename: absoluteFilePath("modules/module-a/ModuleA.js"),
+        code: "import ModuleB from '../../modules/module-b/foo.js'",
+        settings: {
+          ...settings,
+          "boundaries/types": null,
+        },
+      },
+      // Repeat no type provided, check that it continues working
+      {
+        filename: absoluteFilePath("modules/module-a/ModuleA.js"),
+        code: "import ModuleB from '../../modules/module-b/foo.js'",
+        settings: {
+          ...settings,
+          "boundaries/types": null,
+        },
+      },
+      // Helpers cant import another helper if everything is allowed
+      {
+        filename: absoluteFilePath("helpers/helper-a/HelperA.js"),
+        code: "import HelperB from 'helpers/helper-b'",
+        options: [
+          {
+            default: "allow",
+          },
+        ],
+      },
+    ],
+    invalid: [
+      // Helpers can't import another if everything is disallowed
+      {
+        filename: absoluteFilePath("helpers/helper-a/HelperA.js"),
+        code: "import HelperB from 'helpers/helper-b'",
+        options: [
+          {
+            default: "disallow",
+          },
+        ],
+        errors: [
+          {
+            message: errorMessage("helpers", "helpers"),
+            type: "ImportDeclaration",
+          },
+        ],
+      },
+      // Helpers can't import another helper
+      {
+        filename: absoluteFilePath("helpers/helper-a/HelperA.js"),
+        code: "import HelperB from 'helpers/helper-b'",
+        options,
+        errors: [
+          {
+            message: errorMessage("helpers", "helpers"),
+            type: "ImportDeclaration",
+          },
+        ],
+      },
+      // Helpers can't import a component:
+      {
+        filename: absoluteFilePath("helpers/helper-a/HelperA.js"),
+        code: "import ComponentA from 'components/component-a'",
+        options,
+        errors: [
+          {
+            message: errorMessage("helpers", "components"),
+            type: "ImportDeclaration",
+          },
+        ],
+      },
+      // Helpers can't import a module:
+      {
+        filename: absoluteFilePath("helpers/helper-a/HelperA.js"),
+        code: "import ModuleA from 'modules/module-a'",
+        options,
+        errors: [
+          {
+            message: errorMessage("helpers", "modules"),
+            type: "ImportDeclaration",
+          },
+        ],
+      },
+      // Components can't import a module:
+      {
+        filename: absoluteFilePath("components/component-a/ComponentA.js"),
+        code: "import ModuleA from 'modules/module-a'",
+        options,
+        errors: [
+          {
+            message: errorMessage("components", "modules"),
+            type: "ImportDeclaration",
+          },
+        ],
+      },
+    ],
+  });
+};
+
+// deprecated settings
+
+test(SETTINGS.deprecated, [
+  {
+    default: "disallow",
+    rules: [
+      {
+        from: "components",
+        allow: ["helpers", "components"],
+      },
+      {
+        from: "modules",
+        allow: ["helpers", "components", "modules"],
+      },
+    ],
+  },
+]);
+
+// disallow-based options
+
+test(SETTINGS.oneLevel, [
+  {
+    default: "disallow",
+    rules: [
+      {
+        from: "components",
+        allow: ["helpers", "components"],
+      },
+      {
+        from: "modules",
+        allow: ["helpers", "components", "modules"],
+      },
+    ],
+  },
+]);
+
+// allow-based options
+test(SETTINGS.oneLevel, [
+  {
+    default: "allow",
+    rules: [
+      {
+        from: "helpers",
+        disallow: ["modules", "components", "helpers"],
+      },
+      {
+        from: "components",
+        disallow: ["modules"],
+      },
+    ],
+  },
+]);
