+++ test/src/rules/one-level/invalid-settings.spec.js
@@ -1,0 +1,160 @@
+const { ELEMENT_TYPES: RULE } = require("../../../../src/constants/rules");
+const { SETTINGS, createRuleTester, pathResolvers } = require("../../helpers");
+
+const rule = require(`../../../../src/rules/${RULE}`);
+
+const { absoluteFilePath } = pathResolvers("one-level");
+
+const errorMessage = (fileType, dependencyType) =>
+  `Usage of '${dependencyType}' is not allowed in '${fileType}'`;
+
+const test = (settings, options) => {
+  const ruleTester = createRuleTester(settings);
+
+  ruleTester.run(RULE, rule, {
+    // Everything is valid, as settings are wrong
+    valid: [
+      // Helpers can't import another if everything is disallowed
+      {
+        filename: absoluteFilePath("helpers/helper-a/HelperA.js"),
+        code: "import HelperB from 'helpers/helper-b'",
+        options: [
+          {
+            default: "disallow",
+          },
+        ],
+        errors: [
+          {
+            message: errorMessage("helpers", "helpers"),
+            type: "ImportDeclaration",
+          },
+        ],
+      },
+      // Helpers can't import another helper
+      {
+        filename: absoluteFilePath("helpers/helper-a/HelperA.js"),
+        code: "import HelperB from 'helpers/helper-b'",
+        options,
+        errors: [
+          {
+            message: errorMessage("helpers", "helpers"),
+            type: "ImportDeclaration",
+          },
+        ],
+      },
+      // Helpers can't import a component:
+      {
+        filename: absoluteFilePath("helpers/helper-a/HelperA.js"),
+        code: "import ComponentA from 'components/component-a'",
+        options,
+        errors: [
+          {
+            message: errorMessage("helpers", "components"),
+            type: "ImportDeclaration",
+          },
+        ],
+      },
+      // Helpers can't import a module:
+      {
+        filename: absoluteFilePath("helpers/helper-a/HelperA.js"),
+        code: "import ModuleA from 'modules/module-a'",
+        options,
+        errors: [
+          {
+            message: errorMessage("helpers", "modules"),
+            type: "ImportDeclaration",
+          },
+        ],
+      },
+      // Components can't import a module:
+      {
+        filename: absoluteFilePath("components/component-a/ComponentA.js"),
+        code: "import ModuleA from 'modules/module-a'",
+        options,
+        errors: [
+          {
+            message: errorMessage("components", "modules"),
+            type: "ImportDeclaration",
+          },
+        ],
+      },
+    ],
+    invalid: [
+      // Helpers can't import another if everything is disallowed
+      {
+        filename: absoluteFilePath("helpers/helper-a/HelperA.js"),
+        code: "import HelperB from 'helpers/helper-b'",
+        settings: SETTINGS.oneLevel,
+        options: [
+          {
+            default: "disallow",
+          },
+        ],
+        errors: [
+          {
+            message: errorMessage("helpers", "helpers"),
+            type: "ImportDeclaration",
+          },
+        ],
+      },
+    ],
+  });
+};
+
+// no element settings
+
+test(
+  {
+    ...SETTINGS.oneLevel,
+    "boundaries/elements": [],
+  },
+  [
+    {
+      default: "disallow",
+    },
+  ]
+);
+
+// no type
+
+test(
+  {
+    ...SETTINGS.oneLevel,
+    "boundaries/elements": [
+      {
+        pattern: "foo/*",
+        capture: ["elementName"],
+      },
+    ],
+  },
+  []
+);
+
+// no valid match
+
+test(
+  {
+    ...SETTINGS.oneLevel,
+    "boundaries/elements": [
+      {
+        match: "foo",
+      },
+    ],
+  },
+  []
+);
+
+// no valid capture
+
+test(
+  {
+    ...SETTINGS.oneLevel,
+    "boundaries/elements": [
+      {
+        pattern: "foo/*",
+        capture: "foo",
+      },
+    ],
+  },
+  []
+);
